Adds Sanguintare/Matrigrass to Clock/Cult Map
After reflecting I want to see if this will help improve the experience of the map by not restricting the rewards. This way, an explorer won't have to worry about not getting their preferred version of the map for the reward.

Also added a third magical SMES to the NT Outpost section to help prolong power.